requirements.txt
aeval==0.1.3
discord.py==1.7.3
aiohttp==3.7.4.post0
requests==2.28.1
colorama==0.4.5